# frontend/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Face Recognition Attendance - Backend API (inferred)
  version: 1.0.0
  description: |
    OpenAPI spec inferred from the frontend. Use this as a starting point for backend implementation.
servers:
  - url: http://localhost:8000
    description: Local development server (set `VITE_API_BASE` in frontend)
paths:
...
  /api/attendance/mark:
    post:
      summary: Mark attendance (image or reference)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                image_url: { type: string }
                face_id: { type: string }
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Attendance marked
          content:
            application/json:
              schema:
                type: object
                properties:
                  marked: { type: boolean }
                  record:
                    type: object
                    properties:
                      id: { type: string }
                      userId: { type: string }
                      name: { type: string }
                      status: { type: string }
                      timestamp: { type: string }
              example:
                marked: true
                record:
                  id: "rec_1"
                  userId: "EMP001"
                  name: "Alice"
                  status: "Present"
                  timestamp: "2025-11-25T09:00:00Z"
...
  /api/attendance/records:
    get:
      summary: List attendance records (query by date, user, page)
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
        - in: query
          name: userId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        '200':
          description: Records
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        userId: { type: string }
                        name: { type: string }
                        department: { type: string }
                        timestamp: { type: string }
                        status: { type: string }
              example:
                records:
                  - id: "rec_1"
                    userId: "EMP001"
                    name: "Alice"
                    department: "Engineering"
                    timestamp: "2025-11-25T09:00:00Z"
                    status: "Present"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...